requirements.txt

flask
requests
supabase
python-telegram-bot==13.15
